requirements.txt
absl-py==2.3.0
asttokens==3.0.0
cloudpickle==3.1.1
comm==0.2.2
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fsspec==2025.5.1
grpcio==1.72.1
gym==0.26.2
gym-notices==0.0.8
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jupyterlab_widgets==3.0.15
Markdown==3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
overcooked-ai==1.1.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.1
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
